requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
sentence-transformers==2.2.2
pinecone-client==3.0.0
PyMuPDF==1.23.6
requests==2.31.0
numpy==1.24.3
python-multipart==0.0.6